encoder/vision/track_b/requirements.txt
# Core dependencies - versions from requirements-new.txt (verified working)
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
decord==0.6.0
einops==0.6.1
fastapi==0.115.2
filelock==3.16.1
fsspec==2023.9.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.26.4
packaging==23.0
pillow==11.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
starlette==0.40.0
tokenizers==0.21.4
tqdm==4.62.3
transformers==4.52.4
typing_extensions==4.15.0
urllib3==2.2.3
psutil==6.1.0
uvicorn==0.32.0
accelerate==1.8.1
sentencepiece==0.1.99
protobuf==3.20.3
qwen-vl-utils==0.0.11
av==15.1.0
opencv-python==4.9.0.80
easydict==1.13
timm==1.0.15
